vertex-tryon/requirements.txt
google-genai>=0.2.0
flask>=3.0.0
flask-cors>=4.0.0
pillow>=10.0.0
google-auth>=2.23.0
